+++ objectbox/example/flutter/event_management_tutorial/event_manager/android/app/src/main/kotlin/com/example/event_manager_objectbox/MainActivity.kt
@@ -1,0 +1,6 @@
+package com.example.event_manager_objectbox
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}